+++ src/main/java/com/culinary/userservice/ingredient/repository/IngredientRepository.java
@@ -3,5 +3,5 @@ package com.culinary.userservice.ingredient.repository;
 import com.culinary.userservice.ingredient.model.Ingredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
 }
