+++ src/main/java/com/culinary/userservice/ingredient/repository/SpecificRepository.java
@@ -3,5 +3,5 @@ package com.culinary.userservice.ingredient.repository;
 import com.culinary.userservice.ingredient.model.Specific;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecificRepository extends JpaRepository<Specific, Integer> {
+public interface SpecificRepository extends JpaRepository<Specific, Long> {
 }
