+++ src/main/java/com/culinary/userservice/recipe/dto/type/DietTypeDTO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class DietTypeDTO {
-    private int dietTypeId;
+    private long dietTypeId;
     private String dietType;
 }
