+++ src/main/java/com/culinary/userservice/recipe/repository/DietTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DietTypeRepository extends JpaRepository<DietType, Integer> {
+public interface DietTypeRepository extends JpaRepository<DietType, Long> {
 
     Optional<DietType> findDietTypeByDietTypeIgnoreCase(String dietType);
 }
