+++ src/main/java/com/culinary/userservice/recipe/repository/GeneralRecipeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface GeneralRecipeRepository extends JpaRepository<GeneralRecipe, Integer> {
+public interface GeneralRecipeRepository extends JpaRepository<GeneralRecipe, Long> {
 
     @Query("SELECT c FROM GeneralRecipe c WHERE LOWER(c.name) LIKE LOWER(CONCAT('%', :name, '%'))")
     List<GeneralRecipe> findByNameRegex(String name);
